previewer/requirements.txt
pydicom
numpy
pillow
opencv-python
scikit-learn